+++ docs/src/app/pages/docs/components/app-components-lazy-img/app-components-lazy-img.tsx
@@ -43,7 +43,7 @@ export class AppComponentsLazyImg {
 <h2 id="app-components-lazy-img-installation">Installation</h2>
 <p>This component could be added to your web application using the following methods.</p>
 <blockquote>
-<p>If you are using our Starter Kit to develop your presentation, no need to worry about this, this component is included, therefore you could skip the &quot;Installation&quot; chapter.</p>
+<p>If you are using our Starter Kit this template is included. You don&#39;t need to install it so therefore you should skip the &quot;Installation&quot; chapter.</p>
 </blockquote>
 <h3 id="app-components-lazy-img-using-from-a-cdn">Using from a CDN</h3>
 <p>It&#39;s recommended to use <a href="https://unpkg.com/">unpkg</a> to use the <a href="https://deckdeckgo.com">DeckDeckGo</a> lazy image component from a CDN. To do so, add the following include script in the main HTML file of your project:</p>
@@ -108,6 +108,18 @@ export class AppComponentsLazyImg {
 <td>string</td>
 <td></td>
 <td>The image alternate text</td>
+</tr>
+<tr>
+<td>img-width</td>
+<td>number</td>
+<td></td>
+<td>The image width</td>
+</tr>
+<tr>
+<td>img-height</td>
+<td>number</td>
+<td></td>
+<td>The image height</td>
 </tr>
 <tr>
 <td>svg-src</td>
@@ -214,6 +226,21 @@ export class AppComponentsLazyImg {
 <td>--deckgo-lazy-img-object-fit</td>
 <td></td>
 <td>The property object-fit of the image</td>
+</tr>
+<tr>
+<td>--deckgo-lazy-img-opacity-not-loaded</td>
+<td>0</td>
+<td>The opacity of the image when not loaded</td>
+</tr>
+<tr>
+<td>--deckgo-lazy-img-opacity-loaded</td>
+<td>1</td>
+<td>The opacity of the image when loaded</td>
+</tr>
+<tr>
+<td>--deckgo-lazy-img-transition</td>
+<td>opacity 0.15s linear</td>
+<td>The animation of the image, notably use to display smoothly the image when loaded</td>
 </tr>
 </tbody></table>
 <h3 id="app-components-lazy-img-methods">Methods</h3>