+++ docs/src/app/pages/docs/components/app-components-qrcode/app-components-qrcode.tsx
@@ -62,7 +62,7 @@ export class AppComponentsQRCode {
 <h2 id="app-components-qrcode-installation">Installation</h2>
 <p>This component could be added to your web application using the following methods.</p>
 <blockquote>
-<p>If you are using our Starter Kit to develop your presentation, no need to worry about this, this component is included, therefore you could skip the &quot;Installation&quot; chapter.</p>
+<p>If you are using our Starter Kit this template is included. You don&#39;t need to install it so therefore you should skip the &quot;Installation&quot; chapter.</p>
 </blockquote>
 <h3 id="app-components-qrcode-using-deckdeckgo-qr-code-from-a-cdn">Using DeckDeckGo QR Code from a CDN</h3>
 <p>It&#39;s recommended to use <a href="https://unpkg.com/">unpkg</a> to use the <a href="https://deckdeckgo.com">DeckDeckGo</a> Code from a CDN. To do so, add the following include script in the main HTML file of your project:</p>