+++ webcomponents/lazy-img/src/components/lazy-img/deckdeckgo-lazy-img.tsx
@@ -43,6 +43,12 @@ export class DeckdeckgoLazyImg {
   @Prop()
   intrinsicsize: string;
 
+  @Prop()
+  imgWidth: number;
+
+  @Prop()
+  imgHeight: number;
+
   @State()
   private svgContent: string;
 
@@ -189,12 +195,14 @@ export class DeckdeckgoLazyImg {
   }
 
   render() {
+    const hostClass: string = this.imgLoaded ? 'loaded' : '';
+
     if (this.svgContent) {
-      return <Host>
+      return <Host class={hostClass}>
         <div innerHTML={this.svgContent}></div>
       </Host>
     } else {
-      return <Host>
+      return <Host class={hostClass}>
         {this.renderImage()}
       </Host>;
     }
@@ -203,6 +211,7 @@ export class DeckdeckgoLazyImg {
   private renderImage() {
     // @ts-ignore
     return <img alt={this.imgLoaded ? (this.imgAlt ? this.imgAlt : this.imgSrc) : ''} loading="lazy" sizes={this.imgSizes ? this.imgSizes : undefined} intrinsicsize={this.intrinsicsize}
+                width={this.imgWidth} height={this.imgHeight}
                 onLoad={() => this.imgLoaded = true} onError={() => this.loadError()}/>
   }
 }