+++ webcomponents/lazy-img/CHANGELOG.md
@@ -1,3 +1,17 @@
+<a name="1.0.0-rc.2-1"></a>
+# 1.0.0-rc.2-1 (2019-11-29)
+
+### Features
+
+* add new optional properties `img-width` and `img-height` (useful for native lazy loading in Chrome)
+
+<a name="1.0.0-rc.2"></a>
+# 1.0.0-rc.2 (2019-11-29)
+
+### Features
+
+* display smoothly the images once loaded ([#508](https://github.com/deckgo/deckdeckgo/issues/508))
+
 <a name="1.0.0-rc.1"></a>
 # 1.0.0-rc.1 (2019-08-30)